+++ src/Kris/LaravelFormBuilder/Form.php
@@ -193,6 +193,31 @@ class Form
     }
 
     /**
+     * Renders the rest of the form up until the specified field name
+     *
+     * @param string $field_name
+     * @param bool   $showFormEnd
+     * @param bool   $showFields
+     * @return string
+     */
+    public function renderUntil($field_name, $showFormEnd = true, $showFields = true)
+    {
+        $fields = $this->getUnrenderedFields();
+
+        $i = 1;
+        foreach ($fields as $key => $value) {
+            if ($value->getName() == $field_name) {
+                break;
+            }
+            $i++;
+        }
+
+        $fields = array_slice($fields, 0, $i, true);
+
+        return $this->render([], $fields, false, $showFields, $showFormEnd);
+    }
+
+    /**
      * Get single field instance from form object
      *
      * @param $name
@@ -363,6 +388,8 @@ class Form
     {
         $this->model = $model;
 
+        $this->setupNamedModel();
+
         // Rebuild so new data is bound to the fields
         $this->rebuildFields();
 
@@ -454,8 +481,12 @@ class Form
      * @param null   $default
      * @return mixed
      */
-    public function getData($name, $default = null)
+    public function getData($name = null, $default = null)
     {
+        if (is_null($name)) {
+            return $this->data;
+        }
+
         return array_get($this->data, $name, $default);
     }